Fix data races around app data reverting
The issue is happening due to a wrong work with
apps protected data. As we are working with maps
which are containing pointers to objects, we can't
just make a local copy of such maps from the
protected area and feel free to work with them
without an accessor. This is because a local copy
contains pointers to the same objects which should
be accessed through the protected methods only.
To avoid that, accessor lifetime has been increased
while we are working with the map content.
Additionally, mutex have been made recursive for all
maps as sometimes SDL need to iterate over the
protected map and access to the same map in another
methods.