Fix wrong revert of GP and Commands
There was found two issues:
1. Commands revert has been done by internal id
instead of cmdID which makes resumption data
revert sensitive to commands order. This peace
of code was updated to work properly with cmdID
despite of commands resumption order
2. There was a missing part of logic regarding
reset of VRHelp property. Because this property
was not considered during revert, in some cases
SDL does not trigger sending of SetGP. Missing
check has been added to resolve that issue